+++ src/myoddweb.directorywatcher/utils/Helper/RenamedFileSystemEvent.cs
@@ -1,10 +1,11 @@
 ﻿// Licensed to Florent Guelfucci under one or more agreements.
 // Florent Guelfucci licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
+
 using System.IO;
 using myoddweb.directorywatcher.interfaces;
 
-namespace myoddweb.directorywatcher.utils
+namespace myoddweb.directorywatcher.utils.Helper
 {
   /// <inheritdoc cref="IRenamedFileSystemEvent" />
   internal class RenamedFileSystemEvent : FileSystemEvent, IRenamedFileSystemEvent